refactor: package構成を変更
apiパッケージにgithub search関連の処理がまとまるようにした
