+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/ExampleInstrumentedTest.kt
@@ -8,6 +8,7 @@ import io.ktor.client.request.header
 import io.ktor.client.request.parameter
 import io.ktor.client.statement.HttpResponse
 import io.ktor.client.statement.bodyAsText
+import jp.co.yumemi.android.codecheck.api.RepositorySearchResponse
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.json.Json
 import org.junit.Assert.assertEquals
