+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/SearchViewModel.kt
@@ -5,6 +5,8 @@ package jp.co.yumemi.android.codecheck
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import jp.co.yumemi.android.codecheck.api.RepositoryInfo
+import jp.co.yumemi.android.codecheck.api.SearchClient
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.StateFlow
 import org.koin.android.annotation.KoinViewModel
