+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/adapter/OnItemClickListener.kt
@@ -1,6 +1,6 @@
 package jp.co.yumemi.android.codecheck.adapter
 
-import jp.co.yumemi.android.codecheck.RepositoryInfo
+import jp.co.yumemi.android.codecheck.api.RepositoryInfo
 
 interface OnItemClickListener {
     fun itemClick(item: RepositoryInfo)
